move things around, address feedback
