+++ eurofxref-hist.csv
@@ -1,4 +1,11 @@
 Date,USD,JPY,BGN,CYP,CZK,DKK,EEK,GBP,HUF,LTL,LVL,MTL,PLN,ROL,RON,SEK,SIT,SKK,CHF,ISK,NOK,HRK,RUB,TRL,TRY,AUD,BRL,CAD,CNY,HKD,IDR,ILS,INR,KRW,MXN,MYR,NZD,PHP,SGD,THB,ZAR,
+2024-06-11,1.073,168.46,1.9558,N/A,24.706,7.4584,N/A,0.84198,394.63,N/A,N/A,N/A,4.3413,N/A,4.9768,11.2467,N/A,N/A,0.9622,149.5,11.491,N/A,N/A,N/A,34.74,1.626,5.7464,1.4781,7.7839,8.3802,17488.4,3.9923,89.6748,1479.23,19.7008,5.064,1.7513,63.071,1.4517,39.433,19.9574,
+2024-06-10,1.0756,168.74,1.9558,N/A,24.636,7.4592,N/A,0.84565,394.15,N/A,N/A,N/A,4.324,N/A,4.9769,11.333,N/A,N/A,0.9637,149.7,11.512,N/A,N/A,N/A,34.8697,1.6304,5.72,1.4801,7.7957,8.4043,17529.81,4.0242,89.8004,1479.33,19.7531,5.0779,1.7595,63.243,1.4548,39.544,20.1228,
+2024-06-07,1.0898,169.52,1.9558,N/A,24.577,7.4613,N/A,0.8512,388.92,N/A,N/A,N/A,4.288,N/A,4.9763,11.3075,N/A,N/A,0.9696,149.9,11.4865,N/A,N/A,N/A,35.165,1.6345,5.7158,1.4902,7.8916,8.5095,17646.42,4.0723,90.889,1489.97,19.4943,5.1123,1.7596,63.786,1.4653,39.756,20.4821,
+2024-06-06,1.0865,169.74,1.9558,N/A,24.632,7.4592,N/A,0.85088,390.68,N/A,N/A,N/A,4.3015,N/A,4.9758,11.293,N/A,N/A,0.9687,149.7,11.5,N/A,N/A,N/A,35.0263,1.6368,5.7623,1.4879,7.8715,8.4853,17670.67,4.0503,90.7135,1486.44,19.1254,5.1,1.7584,63.734,1.4648,39.679,20.6281,
+2024-06-05,1.0872,169.72,1.9558,N/A,24.654,7.4592,N/A,0.85048,391.73,N/A,N/A,N/A,4.3045,N/A,4.9755,11.3275,N/A,N/A,0.9704,149.5,11.475,N/A,N/A,N/A,35.2316,1.6353,5.7259,1.4867,7.8795,8.4904,17762.62,4.0302,90.708,1491.34,19.1919,5.1077,1.7577,63.938,1.4657,39.906,20.5868,
+2024-06-04,1.0865,168.29,1.9558,N/A,24.755,7.459,N/A,0.85143,392.28,N/A,N/A,N/A,4.3005,N/A,4.975,11.3755,N/A,N/A,0.9703,149.9,11.489,N/A,N/A,N/A,35.2217,1.6364,5.7284,1.486,7.863,8.4901,17680.4,4.005,90.6925,1495.71,19.5326,5.1087,1.7624,63.854,1.464,39.733,20.2986,
+2024-06-03,1.0842,170.09,1.9558,N/A,24.724,7.459,N/A,0.85175,391.4,N/A,N/A,N/A,4.2815,N/A,4.9762,11.4035,N/A,N/A,0.9772,149.5,11.3885,N/A,N/A,N/A,34.9104,1.629,5.6957,1.4784,7.8562,8.4797,17595.37,3.9692,90.125,1493.24,19.0161,5.1033,1.7614,63.653,1.4638,39.937,20.2534,
 2024-05-31,1.0852,170.52,1.9558,N/A,24.709,7.4588,N/A,0.85365,388.83,N/A,N/A,N/A,4.2645,N/A,4.9767,11.421,N/A,N/A,0.9818,149.1,11.383,N/A,N/A,N/A,34.9691,1.6315,5.6418,1.4804,7.8577,8.4838,17641.72,4.0342,90.5255,1501.11,18.4387,5.108,1.7696,63.604,1.4663,39.925,20.2927,
 2024-05-30,1.0815,169.5,1.9558,N/A,24.77,7.4595,N/A,0.85105,389.45,N/A,N/A,N/A,4.2848,N/A,4.9764,11.4995,N/A,N/A,0.9808,148.9,11.4065,N/A,N/A,N/A,34.9038,1.6341,5.6279,1.4828,7.8261,8.4556,17563.67,4.0247,90.073,1488.66,18.4071,5.0874,1.7706,63.354,1.4612,39.745,20.1501,
 2024-05-29,1.0857,170.74,1.9558,N/A,24.731,7.4607,N/A,0.8513,387.08,N/A,N/A,N/A,4.264,N/A,4.9768,11.5215,N/A,N/A,0.9907,149.1,11.417,N/A,N/A,N/A,34.9889,1.637,5.6149,1.4838,7.8706,8.4811,17580.25,4.0153,90.529,1484.62,18.2588,5.1066,1.7716,63.481,1.4652,39.948,19.9,
