+++ pyfifotax/data_structures_event.py
@@ -1,85 +1,23 @@
-import math
 import datetime
 import decimal
 from datetime import datetime
 from decimal import Decimal
 
-from utils import to_decimal, round_decimal
-
-
-# base class for handling curency-valued items
-# e.g. for keeping track of tax payments or
-# bookkeeping of other monetary items
-class Forex:
-    def __init__(self, currency: str, date: datetime, amount: Decimal, comment: str):
-        self.currency = currency
-        self.date = date
-        self.amount = amount
-        # after conversion into domestic currency with
-        # daily or monthly average currency exchange rate
-        self.amount_eur_daily = None
-        self.amount_eur_monthly = None
-        self.comment = comment
-
-    def __repr__(self):
-        return f"{self.currency}(Date: {self.date}, Amount: {self.amount:.2f})"
-
-
-# base class representing an arbitrary asset subject to FIFO treatment
-class FIFOObject:
-    def __init__(
-        self,
-        symbol: str,
-        quantity: Decimal,
-        buy_date: datetime,
-        buy_price: Decimal,
-        currency: str,
-    ):
-        self.symbol = symbol
-        self.currency = currency
-        self.quantity = quantity
-        self.buy_date = buy_date
-        self.sell_date = None
-        self.buy_price = buy_price
-        self.buy_price_eur_daily = None
-        self.buy_price_eur_monthly = None
-        self.sell_price = None
-        self.sell_price_eur_daily = None
-        self.sell_price_eur_monthly = None
-        self.gain_eur_daily = None
-        self.gain_eur_monthly = None
-
-    def total_buy_value(self):
-        return self.quantity * self.buy_price
-
-    def toal_sell_value(self):
-        return self.quantity * self.sell_price
-
-
-# class representing a FOREX object subject to FIFO treatment
-class FIFOForex(FIFOObject):
-    def __init__(
-        self, currency: str, quantity: Decimal, buy_date: datetime, source: str
-    ):
-        # "money" is always a single unit "money"
-        super().__init__(currency, quantity, buy_date, 1, currency)
-        self.source = source  # e.g. "sale", "deposit", "dividend", etc.
-
-    def __repr__(self):
-        return f"{self.symbol}(Quantity: {self.quantity:.2f}, Buy-Date: {self.buy_date:%Y-%b-%d})"
-
-
-# class representing a share subject to FIFO treatment
-class FIFOShare(FIFOObject):
-    def __init__(self, symbol, quantity, buy_date, buy_price, currency):
-        super().__init__(symbol, quantity, buy_date, buy_price, currency)
-
-    def __repr__(self):
-        return f"{self.symbol}(Quantity: {self.quantity}, Buy-Date: {self.buy_date:%Y-%b-%d}, Buy-Price: {self.buy_price:.2f} {self.currency})"
+from pyfifotax.data_structures_fifo import FIFOForex, FIFOShare, Forex
+from pyfifotax.data_structures_dataframe import (
+    ESPPRow,
+    RSURow,
+    BuyOrderRow,
+    SellOrderRow,
+    DividendRow,
+    CurrencyConversionRow,
+    StockSplitRow,
+)
+from pyfifotax.utils import to_decimal, round_decimal
 
 
 class ReportEvent:
-    def __init__(self, date: datetime):
+    def __init__(self, date: datetime, priority: int):
         self.date = date
         # add priority to make sure that buy/deposit transactions
         # are processed before sell transactions
@@ -88,17 +26,17 @@ class ReportEvent:
         # sell: prio 2
         # stocksplit: 3 (at the end as assumed after market-close)
         # currency conversions: 4
-        self.priority = 0
+        self.priority = priority
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
         raise NotImplementedError
 
     @classmethod
     def from_report(cls, df):
         events = []
         for _, row in df.iterrows():
-            events.append(cls.from_report_row(row))
+            events.append(cls.from_df_row(row))
         return events
 
     def __repr__(self) -> str:
@@ -113,14 +51,14 @@ class RSUEvent(ReportEvent):
         received_shares: FIFOShare,
         withheld_shares: FIFOShare,
     ):
-        self.date = date
+        super().__init__(date, 0)
         self.symbol = symbol
         self.received_shares = received_shares
         self.withheld_shares = withheld_shares
-        self.priority = 0
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = RSURow.from_df_row(row)
         recv_share = FIFOShare(
             buy_date=row.date,
             symbol=row.symbol,
@@ -154,20 +92,26 @@ class DividendEvent(ReportEvent):
         received_net_dividend: FIFOForex,
         withheld_tax: Forex,
     ):
-        self.date = date
+        super().__init__(date, 0)
         self.currency = currency
         self.received_dividend = received_dividend
         self.received_net_dividend = received_net_dividend
         self.withheld_tax = withheld_tax
-        self.priority = 0
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = DividendRow.from_df_row(row)
         gross_amount = to_decimal(row.amount)
         tax_amount = to_decimal(row.tax_withholding)
         net_amount = gross_amount - tax_amount
-        assert 0 <= tax_amount <= gross_amount, "Expected non-negative Tax Withholding!"
-        assert 0 <= net_amount <= gross_amount, "Expected non-negative Net Amount"
+        if gross_amount < 0:
+            raise ValueError(
+                f"Expected Capital Gains to be non-negative but got {net_amount}."
+            )
+        if not (0 <= tax_amount <= gross_amount):
+            raise ValueError(
+                f"Expected Tax Withholding {tax_amount} to be non-negative and smaller than the capital gains {gross_amount}."
+            )
         div = Forex(
             currency=row.currency,
             date=row.date,
@@ -199,17 +143,17 @@ class ESPPEvent(ReportEvent):
         contribution: decimal,
         bonus: decimal,
     ):
-        self.date = date
+        super().__init__(date, 0)
         self.symbol = symbol
         self.currency = currency
-        self.priority = 0
 
         self.contribution = contribution
         self.bonus = bonus
         self.received_shares = received_shares
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = ESPPRow.from_df_row(row)
         quantity = to_decimal(row.quantity)
         buy_price = to_decimal(row.buy_price)
         contribution = buy_price * quantity
@@ -246,16 +190,16 @@ class BuyEvent(ReportEvent):
         paid_fees: Forex,
         currency: str,
     ):
-        self.date = date
+        super().__init__(date, 1)
         self.symbol = symbol
         self.received_shares = received_shares
         self.cost_of_shares = cost_of_shares
         self.paid_fees = paid_fees
         self.currency = currency
-        self.priority = 1
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = BuyOrderRow.from_df_row(row)
         buy_price = to_decimal(row.buy_price)
         quantity = to_decimal(row.quantity)
         fees = to_decimal(row.fees)
@@ -301,17 +245,17 @@ class SellEvent(ReportEvent):
         received_forex: FIFOForex,
         paid_fees: Forex,
     ):
-        self.date = date
+        super().__init__(date, 2)
         self.symbol = symbol
         self.currency = currency
         self.quantity = quantity
         self.sell_price = sell_price
         self.received_forex = received_forex
         self.paid_fees = paid_fees
-        self.priority = 2
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = SellOrderRow.from_df_row(row)
         sell_price = to_decimal(row.sell_price)
         quantity = to_decimal(row.quantity)
         fees = to_decimal(row.fees)
@@ -321,9 +265,10 @@ class SellEvent(ReportEvent):
         # TODO: check if rounding mode matches schwab
         net_proceeds = sell_price * quantity - fees
         net_proceeds = round_decimal(net_proceeds)
-        assert (
-            net_proceeds >= 0
-        ), "Expected non-negative net proceeds from sale of shares!"
+        if not (net_proceeds >= 0):
+            raise ValueError(
+                f"Expected non-negative net proceeds from sale of shares but got {net_proceeds}"
+            )
         new_forex = FIFOForex(
             currency=row.currency,
             quantity=net_proceeds,
@@ -351,15 +296,15 @@ class CurrencyConversionEvent(ReportEvent):
         source_currency: str,
         target_currency: str,
     ):
-        self.date = date
+        super().__init__(date, 4)
         self.foreign_amount = foreign_amount
         self.source_fees = source_fees
         self.source_currency = source_currency
         self.target_currency = target_currency
-        self.priority = 4
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = CurrencyConversionRow.from_df_row(row)
         source_fees = Forex(
             currency=row.source_currency,
             date=row.date,
@@ -377,132 +322,15 @@ class CurrencyConversionEvent(ReportEvent):
 
 class StockSplitEvent(ReportEvent):
     def __init__(self, date: datetime, symbol: str, shares_after_split: Decimal):
-        self.date = date
+        super().__init__(date, 3)
         self.symbol = symbol
         self.shares_after_split = shares_after_split
-        self.priority = 3
 
     @staticmethod
-    def from_report_row(row):
+    def from_df_row(row):
+        row = StockSplitRow.from_df_row(row)
         return StockSplitEvent(
             date=row.date,
             symbol=row.symbol,
             shares_after_split=to_decimal(row.shares_after_split),
         )
-
-
-class FIFOQueue:
-    def __init__(self):
-        self.assets: list[FIFOObject] = []
-        self.total_quantity: decimal = to_decimal(0)
-
-    def apply_split(self, shares_after_split: decimal):
-        self.total_quantity = to_decimal(0)
-        for asset in self.assets:
-            asset.quantity = asset.quantity * shares_after_split
-            asset.buy_price = asset.buy_price / shares_after_split
-            self.total_quantity = self.total_quantity + asset.quantity
-
-    def push(self, asset: FIFOObject):
-        if self.is_empty():
-            self.assets = [asset]
-        else:
-            # insert based on buy date ("first in")
-            idx = 0
-            while (idx < len(self.assets)) and (
-                asset.buy_date > self.assets[idx].buy_date
-            ):
-                idx += 1
-            self.assets.insert(idx, asset)
-        self.total_quantity += asset.quantity
-
-    def is_empty(self):
-        return len(self.assets) == 0
-
-    def peek(self):
-        if self.is_empty():
-            raise ValueError("Cannot peek first element from an empty queue.")
-        return self.assets[0]
-
-    def pop(self, quantity: decimal, sell_price: decimal, sell_date: datetime):
-        if math.isclose(quantity, 0, abs_tol=1e-10):
-            return []
-
-        if quantity < 0:
-            raise ValueError("Cannot sell negative amount of equity")
-
-        if quantity > 0 and self.is_empty():
-            raise ValueError(f"Cannot sell equities because there isn't any owned")
-
-        if (
-            not math.isclose(quantity, self.total_quantity)
-            and quantity > self.total_quantity
-        ):
-            symbol = self.peek().symbol
-            asset_type = self.peek().__class__.__name__
-            if asset_type == "FIFOShare":
-                raise ValueError(
-                    f"Cannot sell more {symbol} shares ({quantity:.2f}) than owned overall ({self.total_quantity:.2f})."
-                )
-            elif asset_type == "FIFOForex":
-                raise ValueError(
-                    f"Cannot convert more {symbol} ({quantity:.2f}) than owned overall ({self.total_quantity:.2f})."
-                )
-            else:
-                raise ValueError(
-                    f"Cannot pop quantity ({quantity:.2f}) larger than all quantities ({self.total_quantity:.2f}) in FIFOQueue!"
-                )
-
-        if self.peek().buy_date > sell_date:
-            # Relying on the fact that "assets" are sorted by date
-            symbol = self.peek().symbol
-            raise ValueError(
-                f"Cannot sell the requested {symbol} equity because on the sell transaction date "
-                f"({sell_date.strftime('%Y-%m-%d')}) the requested amount is not available"
-            )
-
-        front_quantity = self.peek().quantity
-        if quantity < front_quantity:
-            pop_asset = from_asset(self.peek(), quantity)
-            pop_asset.sell_price = sell_price
-            pop_asset.sell_date = sell_date
-            self.peek().quantity -= quantity
-            self.total_quantity -= quantity
-            return [pop_asset]
-        elif math.isclose(quantity, front_quantity):
-            self.total_quantity -= front_quantity
-            pop_asset = self.assets.pop(0)
-            pop_asset.sell_date = sell_date
-            pop_asset.sell_price = sell_price
-            return [pop_asset]
-        else:
-            # quantity is larger
-            # pop first item, then call pop on remaining quantity
-            pop_asset = self.assets.pop(0)
-            pop_asset.sell_price = sell_price
-            pop_asset.sell_date = sell_date
-            self.total_quantity -= pop_asset.quantity
-            remaining_quantity = quantity - pop_asset.quantity
-            return [pop_asset] + self.pop(remaining_quantity, sell_price, sell_date)
-
-    def __repr__(self):
-        return self.assets.__repr__()
-
-
-def from_asset(asset, quantity):
-    # intended for usage in "FIFOQueue"
-    if asset.__class__.__name__ == "FIFOForex":
-        new_asset = FIFOForex(asset.currency, quantity, asset.buy_date, asset.source)
-
-    elif asset.__class__.__name__ == "FIFOShare":
-        new_asset = FIFOShare(
-            asset.symbol, quantity, asset.buy_date, asset.buy_price, asset.currency
-        )
-
-    else:
-        asset_type = type(asset)
-        raise ValueError(
-            f"asset is of unsupported type, got {asset_type}, expected 'FIFOForex' or 'FIFOShare'"
-        )
-
-    return new_asset
